Made stuff cleaner for 64bit by using size_t in stead of int in lots of places. Also cleaned up bits of dead wood that this exercise revealed.

+++ include/ambulant/common/video_renderer.h
@@ -102,7 +102,7 @@ class video_renderer : public common::renderer_playable {
 	/// Display video data. Subclass providing this method is responsible for
 	/// eventually free()ing frame. This method is protected because it shares
 	/// the m_lock mutex.
-	virtual void _push_frame(char* frame, int size) = 0;
+	virtual void _push_frame(char* frame, size_t size) = 0;
 
   protected:
 	lib::size m_size;		///< (width, height) of the video data.
