+++ include/ambulant/gui/cocoa/cocoa_dsvideo.h
@@ -56,7 +56,7 @@ class cocoa_dsvideo_renderer :
 	void set_intransition(const lib::transition_info *info) {};
 	void start_outtransition(const lib::transition_info *info) {};
   protected:
-	void _push_frame(char* frame, int size);
+	void _push_frame(char* frame, size_t size);
   private:
 	NSImage *m_image;
 };
