+++ include/ambulant/net/rtsp_datasource.h
@@ -225,7 +225,7 @@ class rtsp_demux : public abstract_demux {
   protected:
 	unsigned long run();
   private:
-	void _push_data_to_sink (int sink_index, timestamp_t pts, const uint8_t*inbuf, int sz);
+	void _push_data_to_sink (int sink_index, timestamp_t pts, const uint8_t*inbuf, size_t sz);
 	static rtsp_context_t *_init_subsessions(rtsp_context_t *context);
 	void _set_position(timestamp_t time);
 	void _cancel();
