+++ src/libambulant/gui/cocoa/cocoa_dsvideo.mm
@@ -136,7 +136,7 @@ my_free_frame(void *ptr, const void *ptr2, size_t size)
 }
 
 void
-cocoa_dsvideo_renderer::_push_frame(char* frame, int size)
+cocoa_dsvideo_renderer::_push_frame(char* frame, size_t size)
 {
 	NSAutoreleasePool *pool = [[NSAutoreleasePool alloc] init];
 	if (m_image) {
@@ -144,7 +144,7 @@ cocoa_dsvideo_renderer::_push_frame(char* frame, int size)
 		m_image = NULL;
 	}
 	AM_DBG lib::logger::get_logger()->debug("cocoa_dsvideo_renderer::_push_frame: size=%d, w*h*3=%d", size, m_size.w * m_size.h * 4);
-	assert(size == (int)(m_size.w * m_size.h * pixel_info_bpp));
+	assert(size == (m_size.w * m_size.h * pixel_info_bpp));
 	// XXXX Who keeps reference to frame?
 	NSSize nssize = NSMakeSize(m_size.w, m_size.h);
 	m_image = [[NSImage alloc] initWithSize: nssize];
