+++ src/libambulant/gui/qt/qt_video_renderer.cpp
@@ -83,17 +83,17 @@ qt_video_renderer::~qt_video_renderer()
 }
 
 void
-qt_video_renderer::_push_frame(char* frame, int size)
+qt_video_renderer::_push_frame(char* frame, size_t size)
 {
 	assert(frame);
-	assert(size == (int)(m_size.w*m_size.h*4));
+	assert(size == (m_size.w*m_size.h*4));
 
 	if (m_image) delete m_image;
 	if (m_data) free(m_data);
 	m_data = (uchar*)frame;
 	m_image = new QImage(m_data,  m_size.w, m_size.h, 32, NULL, 0, QImage::IgnoreEndian);
 
-	AM_DBG lib::logger::get_logger()->debug("qt_video_renderer::_push_frame(0x%x): frame=0x%x, size=%d, m_image=0x%x", (void*) this, (void*)frame, size, m_image);
+	AM_DBG lib::logger::get_logger()->debug("qt_video_renderer::_push_frame(0x%x): frame=0x%x, size=%d, m_image=0x%x", (void*) this, (void*)frame, (int)size, m_image);
 
 }
